requirements.txt
numpy>=1.21.0
networkx>=2.6.3
pycryptodome>=3.10.1
matplotlib>=3.4.3
tqdm>=4.62.0